были написаны модульные тесты. улучшена обработка ошибок. написан класс StringEnum

+++ control_n3/bubbleSelection.cpp
@@ -6,7 +6,7 @@
 using namespace sort;
 
 void sortRowByBubble(vector<int>& row, pair<int, int>& counts) {
-    int n = row.size();
+    int n = static_cast<int> (row.size());
     for (int i = 0; i < n - 1; i++) {
         for (int j = 0; j < n - i - 1; j++) {
             if (row[j] > row[j + 1]) {
